Fixed compilation, broken linking

+++ modules/ftp-client/ftp-client.xq.src/ftpparse.c
@@ -69,7 +69,7 @@ static void initnow(void)
   long year;
 
   initbase();
-  now = time((time_t *) 0) - base;
+  now = (long)(time((time_t *) 0) - base);
 
   if (flagneedcurrentyear) {
     day = now / 86400;
